Add cheesy hack to make the mysql barclamp deploy mysql properly.
With the transition to using InnoDB as the default table type, we need
to be sure to wipe out the innodb logs when we restart the database.

We do this because the deb package defaults to a 5 MB logfile, and the
chef recipe defaults to a 10MB logfile, and the innodb engine dies if
there is any discrepancy instead of resizing the logs.

We should really just nuke the logs when we detect that the logfile size
is going to change, but this will do for now.